+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 1.4.7
 * added `array_has_only_numeric_keys()` function;
 * added `Filesystem::instance()` static method;
+* `Filesystem::createTmpFile()` now throws a `RuntimeException` on failure;
 * `Exceptionist::isWritable()` now throws the `NotWritableException` correctly;
 * extensive improvement of function descriptions and tags.